fixed problems with line head movement
when adding fractional coordinates to rect.x or rect.y, had to get 10
pixels away so it would move a full pixel to the right, otherwise it
just floors out the extra bit of decimal. This meant that subtracting a
fraction from rect.x or rect.y would move the line head, so you didn't
need to be 10 away. the line with math.ceil() accounts for this, and the
extra bits of the if statements prevent the oscillation this created
when not moving in the primary direction. Next, need to look into
cleaning up the update method, it's pretty gross right now.