smooth tracing
line head now snaps to intersections, maybe tweak some of the numbers to
make it smooth, but movement feels pretty good right now. As usual, some
variable renaming and reorganizing of stuff as well, namely passing a
maze object to the trace constructor instead of passing the mazes wall
list to the update function every frame.