revamped line movement
Everything is functionally the same as before, but update() should be a
lot more efficient because it uses math instead of collision checking to
tell when/how the trace should move. Collision checking remains in
trymove to account for future barriers between squares (a mechanic in
the game) and to stop the trace at the borders of the maze. This
movement system should be easier to work with once I get around to
actually drawing a line behind the trace. Oh, and it also fixes that
problem of the line head sometimes being able to teleport through
squares.