big rearrangement into organizedmain
organized is a bit of an overstatement, stuff's really janky and I need
to think about how exactly to structure all my objects before I
continue. I do have clicking and dragging to draw a line though, sorta.
I did and it shouldn't be hard to do but I was working on fading away
after releasing. I'll figure it out.