preparation for constructing borders around ends
small one this time - made a Maze.sortEnds function that generates a
list of end positions for the 4 directions, sorted in the right order
for ease of use in Maze.generateWalls. Time to mull over how to iterate
through these and generate gaps in the maze borders correctly, and how
to round the end nub + deal with collisions with the trace head.