oh my god, I did it
this was the hardest piece of programming I have completed in my career
thus far. That's probably more telling of my lack of experience than of
how hard it was, but I don't want to understate what a pain in the butt
this was to code, but for good reason I suppose, since this update
provides core line movement functionality. To do so, I revamped the way
nullzone barrier values are stored and modified the Trace.tryMove
function to record passing over intersections correctly, which sounded
simple but ended up being much more complicated due to the way line
movement works. I also modified Trace.draw() to include the line body,
which was easy thanks to the pygame.draw.lines method. The lines appear
to be slightly offcenter, however, will fix this later. Anyway, this is
a huge step forward in terms of this project having actual
functionality, so I'm stoked to move forward after this. I definitely
need a break now, though.