Added start circle collision
This turned out to be even easier than expected due to the way I've
implemented barriers for the trace to collide with, turns out I can just
initialize barriers for the starting circle when the trace is created
that, because of the way the Trace's current barrier changing works,
only go into effect when the Trace collides with the start circle after
looping around. Also finally switched around the 2D matrix coordinate
system for maze barrier list that had been bugging me for a while, as
well as removed older versions of files. I don't think I ever really
needed to keep them, with GitHub working the way it does.