beginning compartmentalization algorithm
Added code to convert the trace path to a list of vectors with a start
point and a direction which extends to the next point in that direction.
This is more useful for the maze compartmentalization algorithm, and I
use this list in the preliminary compartmentalizeGrid function, which as
of now puts compartment values in the maze grid based on the squares it
passes. This current algorithm, with the addition of code to "fill in
the blanks" left in between the trace path, would be sufficient were the
player not allowed to start a line anywhere but on the edges of the
maze. As such, I will need to implement not only filling in the blanks
but combining multi-numbered compartments not broken by the trace. Also,
added color to hexagons.