Added square drawing and antialiasing
added antialiasing to some features, including hexagons, start circle,
and the new square elements. antialiasing other parts of the trace and
circles within the maze grid lines would require more effort because I
would need to change the width of the maze grid lines to be even in
pixels, while currently it is the opposite. Next up is programming
solution checking for squares.