Implemented Star Drawing
Implemented star drawing. This one was a real pain in the rear because
of the proportions of the symbol being kinda janky, I actually just
ended up plugging the dimensions I measured from the game into CAD
software to get the right dimensions for all the points. Also, reverted
antialiasing for the trace start circle because there were problems with
correctly overlapping the circle underneath it. Next up is more involved
testing of squares and stars, and making sure they work together
correctly as they do in the game.